refactor(421): extract http + inference families; relocate provider unit
Step 9 of the codegen/api.py decomposition (#421).  HTTP and Inference move out of execute() into vera/runtime/http.py (register_http, stateless) and vera/runtime/inference.py (register_inference(linker, ops_used, env_vars)).  The LLM provider registry (_ProviderConfig, _PROVIDERS) and HTTP call helper (_call_inference_provider) -- module-level in api.py and used only by inference -- relocate into inference.py, eliminating a would-be circular import.  The request timeout, shared by both net families, is renamed _INFERENCE_TIMEOUT->_HTTP_TIMEOUT and homed in http.py (the lower-level layer); inference.py imports it.  Provider-unit unit tests in test_codegen.py repoint their imports/mock.patch targets, and the #591 UTF-8-hygiene source-location tests in test_runtime_traps.py repoint to http.py/inference.py.  api.py 1,480->1,233 lines (-72% from 4,358).  mypy + ruff clean; 27 http/inference codegen + the #591 hygiene tests + the 22-test gate pass; ch09_http/ch09_inference check clean.
